+++ saic-java-mqtt-gateway/src/main/java/net/heberling/ismart/mqtt/SaicMqttGateway.java
@@ -1,5 +1,6 @@
 package net.heberling.ismart.mqtt;
 
+import static net.heberling.ismart.mqtt.MqttGatewayTopics.CLIMATE_REMOTE_TEMPERATURE;
 import static net.heberling.ismart.mqtt.MqttGatewayTopics.REFRESH_MODE;
 import static net.heberling.ismart.mqtt.MqttGatewayTopics.REFRESH_PERIOD;
 
@@ -55,6 +56,7 @@ import org.bn.annotations.ASN1Sequence;
 import org.bn.coders.IASN1PreparedElement;
 import org.eclipse.paho.client.mqttv3.IMqttClient;
 import org.eclipse.paho.client.mqttv3.IMqttDeliveryToken;
+import org.eclipse.paho.client.mqttv3.IMqttMessageListener;
 import org.eclipse.paho.client.mqttv3.MqttCallback;
 import org.eclipse.paho.client.mqttv3.MqttClient;
 import org.eclipse.paho.client.mqttv3.MqttConnectOptions;
@@ -208,8 +210,6 @@ public class SaicMqttGateway implements Callable<Integer> {
               LOGGER.info("Got message for topic {}: {}", topic, message);
               final Matcher setValueMatcher =
                   Pattern.compile(".*/vehicles/([^/]*)/(.*)/set").matcher(topic);
-              final Matcher configurationValueMatcher =
-                  Pattern.compile(".*/vehicles/([^/]*)/(.*)").matcher(topic);
               if (setValueMatcher.matches()) {
                 new Thread(
                         () -> {
@@ -223,12 +223,6 @@ public class SaicMqttGateway implements Callable<Integer> {
                           }
                         })
                     .start();
-              } else if (configurationValueMatcher.matches()) {
-                VehicleState vehicleState =
-                    getVehicleState(mqttAccountPrefix, configurationValueMatcher.group(1));
-                if (!vehicleState.isComplete()) {
-                  vehicleState.configure(configurationValueMatcher.group(2), message);
-                }
               }
             }
 
@@ -238,8 +232,6 @@ public class SaicMqttGateway implements Callable<Integer> {
 
       client.subscribe(mqttAccountPrefix + "/vehicles/+/+/+/set");
       client.subscribe(mqttAccountPrefix + "/vehicles/+/+/+/+/set");
-      client.subscribe(mqttAccountPrefix + "/vehicles/+/" + REFRESH_MODE);
-      client.subscribe(mqttAccountPrefix + "/vehicles/+/" + REFRESH_PERIOD + "/+");
 
       MessageCoder<MP_UserLoggingInReq> loginRequestMessageCoder =
           new MessageCoder<>(MP_UserLoggingInReq.class);
@@ -291,6 +283,34 @@ public class SaicMqttGateway implements Callable<Integer> {
                             vin,
                             getVehicleState(mqttAccountPrefix, vin.getVin()));
                     vehicleHandlerMap.put(vin.getVin(), handler);
+
+                    IMqttMessageListener configListener =
+                        (topic, message) -> {
+                          LOGGER.info("Got retained config from topic {}: {}", topic, message);
+                          final Matcher configurationValueMatcher =
+                              Pattern.compile(".*/vehicles/([^/]*)/(.*)").matcher(topic);
+                          if (configurationValueMatcher.matches()) {
+                            VehicleState vehicleState =
+                                getVehicleState(
+                                    mqttAccountPrefix, configurationValueMatcher.group(1));
+                            if (!vehicleState.isComplete()) {
+                              vehicleState.configure(configurationValueMatcher.group(2), message);
+                            }
+                          }
+                        };
+                    try {
+                      client.subscribe(
+                          mqttAccountPrefix + "/vehicles/+/" + CLIMATE_REMOTE_TEMPERATURE,
+                          configListener);
+                      client.subscribe(
+                          mqttAccountPrefix + "/vehicles/+/" + REFRESH_MODE, configListener);
+                      client.subscribe(
+                          mqttAccountPrefix + "/vehicles/+/" + REFRESH_PERIOD + "/+",
+                          configListener);
+                    } catch (MqttException e) {
+                      throw new RuntimeException(e);
+                    }
+
                     return handler;
                   })
               .map(