+++ saic-java-mqtt-gateway/src/main/java/net/heberling/ismart/mqtt/carconfig/D90HVACSettings.java
@@ -1,0 +1,19 @@
+package net.heberling.ismart.mqtt.carconfig;
+
+// India "GLOSTER";
+public class D90HVACSettings implements HVACSettings {
+  @Override
+  public Integer getMinAllowedTemp() {
+    return 17;
+  }
+
+  @Override
+  public Integer getMaxAllowedTemp() {
+    return 31;
+  }
+
+  @Override
+  public byte mapTempToSaicApi(Integer temp) {
+    return (byte) normalizeTemperature(temp);
+  }
+}