+++ saic-java-mqtt-gateway/src/main/java/net/heberling/ismart/mqtt/carconfig/ZS11EHVACSettings.java
@@ -1,0 +1,19 @@
+package net.heberling.ismart.mqtt.carconfig;
+
+// Default India
+public class ZS11EHVACSettings implements HVACSettings {
+  @Override
+  public Integer getMinAllowedTemp() {
+    return 16;
+  }
+
+  @Override
+  public Integer getMaxAllowedTemp() {
+    return 30;
+  }
+
+  @Override
+  public byte mapTempToSaicApi(Integer temp) {
+    return (byte) normalizeTemperature(temp);
+  }
+}